OtoLauncher: Fix four bugs
After switch the language, the desktop right menu dialog is still the
previous language, and all functions in menu dialog are fail.

After switch the language, the icons's name about MyComputer and
Recycle is sill the previous language.

The sorting function is missed, because the sorting of the reading
chart is not considerated.

Romove a useless onTouchListener.

Signed-off-by: Wang Zhixu <potatomagic@163.com>
Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>